Support description in FASTA/FASTQ input when creating perfect reads
to_perfect_reads.py does not support input which have a description
text after an identifier in input file. It blindly appends the numbers
to the identifier but if the line contained some description words
separated by spaces from the identifier it will append the values after
the description. This commit ensures the values will be appended to the
FASTA/FASTQ identifier even if it is followed by a description text.